requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=1.4.0=pypi_0
ansi2html=1.8.0=pypi_0
anyio=3.7.0=pypi_0
argon2-cffi=21.3.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
astunparse=1.6.3=pypi_0
attrs=23.1.0=pypi_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=pyhd8ed1ab_3
backports.functools_lru_cache=2.0.0=pyhd8ed1ab_0
beautifulsoup4=4.12.2=pypi_0
blas=2.106=mkl
bleach=6.0.0=pypi_0
bob=11.3.0=pypi_0
bob-extension=7.0.3=pypi_0
bob-io-base=5.0.2=pypi_0
brotli-python=1.0.9=py37hf2a7229_7
ca-certificates=2023.11.17=h56e8100_0
cachetools=5.3.1=pypi_0
certifi=2023.11.17=pyhd8ed1ab_0
chardet=4.0.0=pypi_0
charset-normalizer=3.3.2=pyhd8ed1ab_0
click=8.1.3=pypi_0
click-plugins=1.1.1=pypi_0
cmake=3.26.3=pypi_0
colorama=0.4.6=pyhd8ed1ab_0
configargparse=1.7=pypi_0
cuda-cccl=12.2.140=h57928b3_0
cuda-cccl-impl=2.0.1=h57928b3_1
cuda-cccl_win-64=12.2.140=h57928b3_0
cuda-cudart=11.7.99=0
cuda-cudart-dev=11.7.99=0
cuda-cupti=11.7.101=0
cuda-libraries=11.7.1=0
cuda-libraries-dev=11.7.1=0
cuda-nvrtc=11.7.99=0
cuda-nvrtc-dev=11.7.99=0
cuda-nvtx=11.7.91=0
cuda-runtime=11.7.1=0
cuda-version=12.2=he2b69de_2
cycler=0.11.0=pypi_0
dash=2.11.1=pypi_0
dash-core-components=2.0.0=pypi_0
dash-html-components=2.0.0=pypi_0
dash-table=5.0.0=pypi_0
debugpy=1.6.3=py37hf2a7229_0
decorator=5.1.1=pyhd8ed1ab_0
defusedxml=0.7.1=pypi_0
dlib=19.24.1=pypi_0
entrypoints=0.4=pyhd8ed1ab_0
exceptiongroup=1.1.1=pypi_0
facenet-pytorch=2.5.3=pypi_0
fastjsonschema=2.17.1=pypi_0
flask=2.2.5=pypi_0
flatbuffers=23.5.26=pypi_0
fonttools=4.38.0=pypi_0
freetype=2.12.1=hdaf720e_2
freetype-py=2.4.0=pypi_0
gast=0.4.0=pypi_0
giflib=5.2.1=h64bf75a_3
google-auth=2.19.1=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
google-pasta=0.2.0=pypi_0
grad-cam=1.4.8=pypi_0
grpcio=1.54.2=pypi_0
h5py=3.8.0=pypi_0
idna=2.10=pypi_0
imageio=2.30.0=pypi_0
imageio-ffmpeg=0.4.8=pypi_0
importlib-metadata=6.6.0=pypi_0
importlib-resources=5.12.0=pypi_0
imutils=0.5.4=pypi_0
intel-openmp=2024.0.0=h57928b3_49840
ipykernel=6.16.2=pyh025b116_0
ipython=7.33.0=py37h03978a9_0
ipython-genutils=0.2.0=pypi_0
ipywidgets=8.0.6=pypi_0
itsdangerous=2.1.2=pypi_0
jedi=0.19.1=pyhd8ed1ab_0
jinja2=3.1.2=pypi_0
joblib=1.2.0=pypi_0
jpeg=9e=hcfcfb64_3
jsonpatch=1.33=pypi_0
jsonpointer=2.4=pypi_0
jsonschema=4.17.3=pypi_0
jupyter=1.0.0=pypi_0
jupyter-console=6.6.3=pypi_0
jupyter-core=4.12.0=pypi_0
jupyter-server=1.24.0=pypi_0
jupyter_client=7.4.9=pyhd8ed1ab_0
jupyter_core=4.11.1=py37h03978a9_0
jupyterlab-pygments=0.2.2=pypi_0
jupyterlab-widgets=3.0.7=pypi_0
keras=2.11.0=pypi_0
kiwisolver=1.4.4=pypi_0
lerc=3.0=hd77b12b_0
libblas=3.9.0=6_mkl
libcblas=3.9.0=6_mkl
libclang=16.0.0=pypi_0
libcublas=11.10.3.66=0
libcublas-dev=11.10.3.66=0
libcufft=10.7.2.124=0
libcufft-dev=10.7.2.124=0
libcurand=10.3.3.141=h63175ca_0
libcurand-dev=10.3.3.141=h63175ca_0
libcusolver=11.4.0.1=0
libcusolver-dev=11.4.0.1=0
libcusparse=11.7.4.91=0
libcusparse-dev=11.7.4.91=0
libdeflate=1.17=h2bbff1b_1
liblapack=3.9.0=6_mkl
liblapacke=3.9.0=6_mkl
libnpp=11.7.4.75=0
libnpp-dev=11.7.4.75=0
libnvjpeg=11.8.0.2=0
libnvjpeg-dev=11.8.0.2=0
libpng=1.6.39=h19919ed_0
libsodium=1.0.18=h8d14728_1
libsqlite=3.44.2=hcfcfb64_0
libtiff=4.5.1=hd77b12b_0
libuv=1.44.2=hcfcfb64_1
libwebp=1.2.4=hbc33d0d_1
libwebp-base=1.2.4=h2bbff1b_1
libzlib=1.2.13=hcfcfb64_5
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markdown=3.4.3=pypi_0
markupsafe=2.1.3=pypi_0
matplotlib=3.5.3=pypi_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
megatools=0.0.4=pypi_0
mistune=2.0.5=pypi_0
mkl=2020.4=hb70f87d_311
msys2-conda-epoch=20160418=1
mtcnn=0.1.1=pypi_0
nbclassic=1.0.0=pypi_0
nbclient=0.7.4=pypi_0
nbconvert=7.4.0=pypi_0
nbformat=5.7.0=pypi_0
nest-asyncio=1.5.9=pyhd8ed1ab_0
networkx=2.6.3=pypi_0
ninja=1.11.1.1=pypi_0
notebook=6.5.4=pypi_0
notebook-shim=0.2.3=pypi_0
numpy=1.21.6=py37h2830a78_0
oauthlib=3.2.2=pypi_0
open3d=0.17.0=pypi_0
opencv-contrib-python=4.7.0.72=pypi_0
opencv-python=4.7.0.72=pypi_0
openssl=3.2.0=hcfcfb64_1
opt-einsum=3.3.0=pypi_0
packaging=23.2=pyhd8ed1ab_0
pandas=1.3.5=pypi_0
pandocfilters=1.5.0=pypi_0
parso=0.8.3=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pillow=9.4.0=py37hd77b12b_0
pip=23.3.2=pyhd8ed1ab_0
pkgutil-resolve-name=1.3.10=pypi_0
plotly=5.15.0=pypi_0
prometheus-client=0.17.0=pypi_0
prompt-toolkit=3.0.42=pyha770c72_0
protobuf=3.19.6=pypi_0
psutil=5.9.1=py37hcc03f2d_0
pyasn1=0.5.0=pypi_0
pyasn1-modules=0.3.0=pypi_0
pyglet=2.0.9=pypi_0
pygments=2.17.2=pyhd8ed1ab_0
pyopengl=3.1.0=pypi_0
pyparsing=3.0.9=pypi_0
pyrealsense2=2.54.2.5684=pypi_0
pyrender=0.1.45=pypi_0
pyrsistent=0.19.3=pypi_0
pysocks=1.7.1=py37h03978a9_5
python=3.7.12=h900ac77_100_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python-utils=3.5.2=pypi_0
python-version=0.0.2=pypi_0
python_abi=3.7=4_cp37m
pytorch=1.13.1=py3.7_cuda11.7_cudnn8_0
pytorch-cuda=11.7=h16d0643_5
pytorch-mutex=1.0=cuda
pytz=2023.3=pypi_0
pywavelets=1.3.0=pypi_0
pywin32=303=py37hcc03f2d_0
pywinpty=2.0.10=pypi_0
pyzmq=24.0.1=py37h7347f05_0
qtconsole=5.4.3=pypi_0
qtpy=2.3.1=pypi_0
requests=2.25.1=pypi_0
requests-oauthlib=1.3.1=pypi_0
retrying=1.3.4=pypi_0
rsa=4.9=pypi_0
scikit-image=0.19.3=pypi_0
scikit-learn=1.0.2=pypi_0
scipy=1.7.3=pypi_0
seaborn=0.12.2=pypi_0
send2trash=1.8.2=pypi_0
setuptools=69.0.3=pyhd8ed1ab_0
six=1.16.0=pyh6c4a22f_0
sklearn=0.0.post5=pypi_0
sniffio=1.3.0=pypi_0
soupsieve=2.4.1=pypi_0
sqlite=3.44.2=hcfcfb64_0
tenacity=8.2.2=pypi_0
tensorboard=2.11.2=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
tensorflow=2.11.0=pypi_0
tensorflow-estimator=2.11.0=pypi_0
tensorflow-intel=2.11.0=pypi_0
tensorflow-io-gcs-filesystem=0.31.0=pypi_0
termcolor=2.3.0=pypi_0
terminado=0.17.1=pypi_0
threadpoolctl=3.1.0=pypi_0
tifffile=2021.11.2=pypi_0
tinycss2=1.2.1=pypi_0
tk=8.6.13=h5226925_1
torchattacks=3.4.0=pypi_0
torchaudio=0.13.1=pypi_0
torchinfo=1.8.0=pypi_0
torchnet=0.0.4=pypi_0
torchsummary=1.5.1=pypi_0
torchvision=0.14.1=pypi_0
tornado=6.2=py37hcc03f2d_0
tqdm=4.64.1=py37haa95532_0
traitlets=5.9.0=pyhd8ed1ab_0
trimesh=3.23.5=pypi_0
ttach=0.0.3=pypi_0
typing_extensions=4.5.0=pyha770c72_0
ucrt=10.0.22621.0=h57928b3_0
urllib3=2.1.0=pyhd8ed1ab_0
utils=1.0.2=pypi_0
vc=14.3=hcf57466_18
vc14_runtime=14.38.33130=h82b7239_18
visdom=0.2.4=pypi_0
vs2015_runtime=14.38.33130=hcb4865c_18
wcwidth=0.2.6=pyhd8ed1ab_0
webencodings=0.5.1=pypi_0
websocket-client=1.5.2=pypi_0
werkzeug=2.2.3=pypi_0
wheel=0.42.0=pyhd8ed1ab_0
widgetsnbextension=4.0.7=pypi_0
win_inet_pton=1.1.0=pyhd8ed1ab_6
wrapt=1.15.0=pypi_0
xz=5.4.5=h8cc25b3_0
zeromq=4.3.4=h0e60522_1
zipp=3.15.0=pypi_0
zlib=1.2.13=hcfcfb64_5
zstd=1.5.5=h12be248_0
